+++ mitra_models/src/activitypub/mod.rs
@@ -1,2 +1,3 @@
 pub mod helpers;
 pub mod queries;
+mod types;